Working interactions with DEFAULT
